Move `rank()` to a separate module
Search now only calculates PageRank with respect to other matches, not
the entire vault
